+++ packages/saved-views-react/src/SavedView.ts
@@ -2,22 +2,28 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+import type { ViewData } from "@itwin/saved-views-client";
 import type { ReactNode } from "react";
 
 export interface SavedView {
   id: string;
   displayName: string;
+  viewData?: ViewData | undefined;
   groupId?: string | undefined;
   creatorId?: string | undefined;
   tagIds?: string[] | undefined;
   shared?: boolean | undefined;
   thumbnail?: ReactNode | string | undefined;
-  /** `extensionName` and `data` pairs. */
-  extensions?: Map<string, string> | undefined;
+  extensions?: SavedViewExtension[] | undefined;
   /** Time the saved view was created as an ISO8601 string, `"YYYY-MM-DDTHH:mm:ss.sssZ"` */
-  creationTime?: string;
+  creationTime?: string | undefined;
   /** Time the saved view was last modified as an ISO8601 string, `"YYYY-MM-DDTHH:mm:ss.sssZ"` */
-  lastModified?: string;
+  lastModified?: string | undefined;
+}
+
+export interface SavedViewExtension {
+  extensionName: string;
+  data: string;
 }
 
 export interface SavedViewTag {
@@ -31,3 +37,5 @@ export interface SavedViewGroup {
   creatorId?: string | undefined;
   shared?: boolean | undefined;
 }
+
+export type WriteableSavedViewProperties = Omit<SavedView, "id" | "creatorId" | "creationTime" | "lastModified">;